requirements.txt
# FinRL_Crypto Dependencies
# Install: pip install -r requirements.txt

# Exchange API (unified interface for Binance, Bitget, etc.)
ccxt>=4.0.0

# Binance API (used by processor_Binance.py for historical data download)
python-binance==1.0.19

# Core ML libraries
torch==2.10.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<2.0.0

# Reinforcement Learning
elegantrl==0.3.5
gymnasium>=0.28.0

# Optimization
optuna==3.4.0

# Technical Analysis
TA-Lib>=0.4.25
stockstats==0.5.0

# Visualization
matplotlib==3.7.3
seaborn==0.12.2

# Network
requests>=2.28.0
urllib3>=1.26.0

# Utilities
tqdm>=4.64.0
psutil>=5.9.0
joblib>=1.2.0

# Statistics
scipy>=1.9.0
scikit-learn>=1.1.0

# Timezone
pytz>=2022.1

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.8.0
certifi>=2022.9.0